+++ src/tests/conftest.py
@@ -5,3 +5,5 @@ from tests.fixtures.project import project  # noqa: F401 pylint: disable=unused-
 from tests.fixtures.dataset import dataset  # noqa: F401 pylint: disable=unused-import
 from tests.fixtures.file_manager import file_manager  # noqa: F401 pylint: disable=unused-import
 from tests.fixtures.image import image  # noqa: F401 pylint: disable=unused-import
+from tests.fixtures.label import label  # noqa: F401 pylint: disable=unused-import
+